feat(typstPackages): rename `typstPackages` to `unstableTypstPackages`
Typix support for Typst packages is considered unstable until stability
can be guaranteed upstream. This means that flake builds using this
attribute which were previously successful could start breaking at any
time (even without updating Typst, for example if a tarball is modified
or removed on the registry).

To emphasize the unstable status of this support, the `typstPackages`
attribute will be tentatively renamed to `unstableTypstPackages`.
Eventually, when stability is provided upstream (and this attribute's
API is stabilized), this can be renamed to `typstPackages` (with the old
name kept as an undocumented alias to avoid breaking builds).